index fixes

+++ src/main/java/com/aurd/Student/Model/Request/GetIndexRequest.java
@@ -3,7 +3,7 @@ package com.aurd.Student.Model.Request;
 public class GetIndexRequest {
 
 
-    int  inst_id;
+    Long  inst_id;
 
 
     long studId;
@@ -53,12 +53,11 @@ public class GetIndexRequest {
         this.filter = filter;
     }
 
-
-    public int getInst_id() {
+    public Long getInst_id() {
         return inst_id;
     }
 
-    public void setInst_id(int inst_id) {
+    public void setInst_id(Long inst_id) {
         this.inst_id = inst_id;
     }
 }
